+++ src/main/java/me/efekos/simpler/Simpler.java
@@ -1,12 +1,11 @@
 package me.efekos.simpler;
 
 import me.efekos.simpler.config.MessageConfiguration;
-import org.bukkit.plugin.java.JavaPlugin;
 
 /**
  * Main class of Simpler.
  */
-public final class Simpler extends JavaPlugin {
+public final class Simpler {
     /**
      * Main configuration for messages that Simpler uses.
      */
@@ -26,24 +25,5 @@ public final class Simpler extends JavaPlugin {
      */
     public static void changeMessageConfiguration(MessageConfiguration configuration) {
         Simpler.configuration = configuration;
-    }
-
-    @Override
-    public void onEnable() {
-
-        //this thing simply makes a /home set <pos> command with nms. I'll use it later 👀👀
-        //CommandDispatcher<CommandListenerWrapper> dispatcher = MinecraftServer.getServer().vanillaCommandDispatcher.a();
-
-        //
-        //dispatcher.register(net.minecraft.commands.CommandDispatcher.a("home")
-        //        .then(net.minecraft.commands.CommandDispatcher.a("set")
-        //                .then(net.minecraft.commands.CommandDispatcher.a("pos", ArgumentPosition.a())
-        //                        .executes(commandContext -> {
-        //                            commandContext.getSource().a(IChatBaseComponent.a("sdf"));
-        //                            return 1;
-        //                        })
-        //                )
-        //        )
-        //);
     }
 }